KermitCode_1-Dev06 Code stable !

+++ Dev_Tmp/BinChecks.h
@@ -3,7 +3,7 @@
 #ifndef _BinChecks_h
 #define _BinChecks_h
 
-bool Check_Alarm(int _Index, byte _CurrentHour, byte _CurrentMinute);
+bool Check_Alarm(int _Index, int _CurrentHour, int _CurrentMinute);
 bool Check_Week (int _Index, byte _CurrentWeekNumber );
 bool Check_Month(int _Index, int _CurrentMonth);
 bool Check_BinActive(int _Index);
